+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed get_bin() in offspot_runtime to accomodate in-venv offspot-runtime
 - Country Code `00` for `ap` not accepted anymore (hostapd). Defaults to `US`
 - Dashboard entry now contains `ident` of package as well
+- Now developping on py3.11 as its default on base-image (bookworm)
 
 ## [1.4.5] - 2023-10-24