+++ CHANGELOG.md
@@ -14,9 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fixed typo in JSON catalog
-- Builder now disables kiwix-serve homepage (redirect to fqdn)
 - Fixed installed dnsmasq-spoof service (was not using venv-aware path)
-- Dashboard yaml config size properly reported
+- [builder] Dashboard yaml config size properly reported
+- [builder] now disables kiwix-serve homepage (redirect to fqdn)
+- [builder] enabling captive portal now sets dhcp-range automatically
 
 ## [1.4.6] - 2023-12-07